renderer-pi/requirements.txt
pygame==2.5.2
opencv-python==4.8.1.78
numpy==1.26.2
requests==2.31.0
websocket-client==1.7.0
